requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
python-dotenv
pandas
docx2txt
qdrant-client
langchain-qdrant